+++ src/main/java/com/br_ifg/clientbackend/service/ClienteService.java
@@ -1,25 +1,25 @@
-package com.example.clientbackend.service;
+package com.br_ifg.clientbackend.service;
 
-import com.example.clientbackend.entity.Cliente;
-import com.example.clientbackend.repository.ClienteRepository;
+import com.br_ifg.clientbackend.entity.Cliente;
+import com.br_ifg.clientbackend.repository.ClienteRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ClienteService {
-    private final ClienteRepository repository;
 
-    public ClienteService(ClienteRepository repository) {
-        this.repository = repository;
-    }
+    @Autowired
+    private ClienteRepository repository;
 
     public List<Cliente> findAll() {
         return repository.findAll();
     }
 
-    public Cliente findById(Long id) {
-        return repository.findById(id).orElseThrow(() -> new RuntimeException("Cliente não encontrado"));
+    public Optional<Cliente> findById(Long id) {
+        return repository.findById(id);
     }
 
     public Cliente save(Cliente cliente) {
@@ -27,6 +27,10 @@ public class ClienteService {
     }
 
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Cliente não encontrado");
+        }
         repository.deleteById(id);
     }
 }
+
